+++ lib/ab_admin/config/base.rb
@@ -18,10 +18,9 @@ module AbAdmin
 
       def self.default_for_model(model, options={})
         new.tap do |builder|
-          builder.field(:id) unless options[:skip].try(:include?, :id)
           model.column_names.each do |column|
             column_name = column.to_sym
-            next if options[:skip].try(:include?, column_name)
+            next if column_name == :id || options[:skip].try(:include?, column_name)
             builder.field(column_name)
           end
         end