+++ lib/ab_admin/utils/eval_helpers.rb
@@ -87,13 +87,15 @@ module AbAdmin
       def call_method_or_proc_on(obj, symbol_or_proc, options = {})
         exec = options[:exec].nil? ? true : options[:exec]
         case symbol_or_proc
-          when Symbol, String
-            obj.send(symbol_or_proc.to_sym, *options[:attrs])
+          when String
+            ActiveSupport::Deprecation.warn('`call_method_or_proc_on` don\'t accept method name as String, use Symbol instead') if symbol_or_proc =~ /\A\w+\z/
+            symbol_or_proc
+          when Symbol
+            obj.public_send(symbol_or_proc.to_sym, *options[:attrs])
           when Proc
             exec ? obj.instance_exec(&symbol_or_proc) : symbol_or_proc.call(obj)
         end
       end
-
     end
   end
 end