+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   * removed Asset#format_created_at
   * removed Time#formatted_datetime, Time#formatted_date
   * date picker format changed to `yyyy-mm-dd` 
+  * Deprecated `Hash#no_blank` in favor of `Hash#reject_blank`
 
 ### 0.7.0
 * enhancements